+++ src/main/java/com/gonggoo/gonggoo/common/config/JpaAuditionConfig.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @Configuration
 @EnableJpaAuditing
-public class JpaAuditionConfiguration {
+public class JpaAuditionConfig {
 }
